+++ HISTORY.rst
@@ -4,6 +4,8 @@ History
 
 0.14.x
 ------
+* Fix in `subset.subset_gridpoint` for dask array coordinates.
+* Modified `subset_shape` to support subsetting with GeoPandas datatypes directly.
 * Fix in `subset.wrap_lons_and_split_at_greenwich` to preserve multi-region dataframes.
 * Improve the memory use of `indices.growing_season_length`.
 * Better handling of data with atypically named `lat` and `lon` dimensions.